Rename example recorder files to 'writer'/'rewriter' (#2347)
