+++ rosbag2_examples/rosbag2_examples_cpp/CMakeLists.txt
@@ -24,20 +24,20 @@ find_package(rosbag2_cpp REQUIRED)
 find_package(rosbag2_transport REQUIRED)
 find_package(example_interfaces REQUIRED)
 
-add_executable(simple_bag_recorder src/simple_bag_recorder.cpp)
-target_link_libraries(simple_bag_recorder
+add_executable(simple_bag_writer src/simple_bag_writer.cpp)
+target_link_libraries(simple_bag_writer
   rclcpp::rclcpp
   rosbag2_cpp::rosbag2_cpp
   ${example_interfaces_TARGETS}
 )
 
 install(TARGETS
-  simple_bag_recorder
+  simple_bag_writer
   DESTINATION lib/${PROJECT_NAME}
 )
 
-add_executable(compressed_bag_recorder src/compressed_bag_recorder.cpp)
-target_link_libraries(compressed_bag_recorder
+add_executable(compressed_bag_writer.cpp src/compressed_bag_writer.cpp)
+target_link_libraries(compressed_bag_writer.cpp
   rclcpp::rclcpp
   rosbag2_cpp::rosbag2_cpp
   rosbag2_compression::rosbag2_compression
@@ -45,7 +45,7 @@ target_link_libraries(compressed_bag_recorder
 )
 
 install(TARGETS
-  compressed_bag_recorder
+  compressed_bag_writer.cpp
   DESTINATION lib/${PROJECT_NAME}
 )
