+++ rosbag2_examples/rosbag2_examples_cpp/src/compressed_bag_writer.cpp
@@ -1,4 +1,4 @@
-// Copyright 2021 Open Source Robotics Foundation
+// Copyright 2025 Open Source Robotics Foundation
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,28 +17,38 @@
 #include "example_interfaces/msg/string.hpp"
 #include "rclcpp/rclcpp.hpp"
 
+#include "rosbag2_compression/compression_options.hpp"
+#include "rosbag2_compression/sequential_compression_writer.hpp"
+
 #include "rosbag2_cpp/writer.hpp"
 
 using std::placeholders::_1;
 
-class SimpleBagRecorder : public rclcpp::Node
-{
+class CompressedBagWriter : public rclcpp::Node {
 public:
-  SimpleBagRecorder()
-  : Node("simple_bag_recorder")
+  CompressedBagWriter()
+  : Node("compressed_bag_writer")
   {
-    writer_ = std::make_unique<rosbag2_cpp::Writer>();
+    rosbag2_compression::CompressionOptions compression_options;
 
+    compression_options.compression_format = "zstd";
+    compression_options.compression_mode = rosbag2_compression::CompressionMode::FILE;
+
+    auto compressed_writer = std::make_unique<rosbag2_compression::SequentialCompressionWriter>(
+      compression_options);
+
+    writer_ = std::make_unique<rosbag2_cpp::Writer>(std::move(compressed_writer));
     writer_->open("my_bag");
 
-    subscription_ = create_subscription<example_interfaces::msg::String>(
-      "chatter", 10, std::bind(&SimpleBagRecorder::topic_callback, this, _1));
+    subscription_ = this->create_subscription<example_interfaces::msg::String>(
+      "chatter", 10, std::bind(&CompressedBagWriter::topic_callback, this, _1));
   }
 
 private:
   void topic_callback(std::shared_ptr<const rclcpp::SerializedMessage> msg) const
   {
     rclcpp::Time time_stamp = this->now();
+
     writer_->write(msg, "chatter", "example_interfaces/msg/String", time_stamp);
   }
 
@@ -46,10 +56,10 @@ private:
   std::unique_ptr<rosbag2_cpp::Writer> writer_;
 };
 
-int main(int argc, char * argv[])
+int main(int argc, char *argv[])
 {
   rclcpp::init(argc, argv);
-  rclcpp::spin(std::make_shared<SimpleBagRecorder>());
+  rclcpp::spin(std::make_shared<CompressedBagWriter>());
   rclcpp::shutdown();
   return 0;
 }
