+++ rosbag2_examples/rosbag2_examples_cpp/src/simple_bag_writer.cpp
@@ -1,4 +1,4 @@
-// Copyright 2025 Open Source Robotics Foundation
+// Copyright 2021 Open Source Robotics Foundation
 //
 // Licensed under the Apache License, Version 2.0 (the "License");
 // you may not use this file except in compliance with the License.
@@ -17,38 +17,28 @@
 #include "example_interfaces/msg/string.hpp"
 #include "rclcpp/rclcpp.hpp"
 
-#include "rosbag2_compression/compression_options.hpp"
-#include "rosbag2_compression/sequential_compression_writer.hpp"
-
 #include "rosbag2_cpp/writer.hpp"
 
 using std::placeholders::_1;
 
-class CompressedBagRecorder : public rclcpp::Node {
+class SimpleBagWriter : public rclcpp::Node
+{
 public:
-  CompressedBagRecorder()
-  : Node("compressed_bag_recorder")
+  SimpleBagWriter()
+  : Node("simple_bag_writer")
   {
-    rosbag2_compression::CompressionOptions compression_options;
+    writer_ = std::make_unique<rosbag2_cpp::Writer>();
 
-    compression_options.compression_format = "zstd";
-    compression_options.compression_mode = rosbag2_compression::CompressionMode::FILE;
-
-    auto compressed_writer = std::make_unique<rosbag2_compression::SequentialCompressionWriter>(
-      compression_options);
-
-    writer_ = std::make_unique<rosbag2_cpp::Writer>(std::move(compressed_writer));
     writer_->open("my_bag");
 
-    subscription_ = this->create_subscription<example_interfaces::msg::String>(
-      "chatter", 10, std::bind(&CompressedBagRecorder::topic_callback, this, _1));
+    subscription_ = create_subscription<example_interfaces::msg::String>(
+      "chatter", 10, std::bind(&SimpleBagWriter::topic_callback, this, _1));
   }
 
 private:
   void topic_callback(std::shared_ptr<const rclcpp::SerializedMessage> msg) const
   {
     rclcpp::Time time_stamp = this->now();
-
     writer_->write(msg, "chatter", "example_interfaces/msg/String", time_stamp);
   }
 
@@ -56,10 +46,10 @@ private:
   std::unique_ptr<rosbag2_cpp::Writer> writer_;
 };
 
-int main(int argc, char *argv[])
+int main(int argc, char * argv[])
 {
   rclcpp::init(argc, argv);
-  rclcpp::spin(std::make_shared<CompressedBagRecorder>());
+  rclcpp::spin(std::make_shared<SimpleBagWriter>());
   rclcpp::shutdown();
   return 0;
 }
