+++ rosbag2_examples/rosbag2_examples_py/rosbag2_examples_py/compressed_bag_writer.py
@@ -1,4 +1,4 @@
-# Copyright 2023 Open Source Robotics Foundation, Inc.
+# Copyright 2025 Open Source Robotics Foundation, Inc.
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -11,6 +11,7 @@
 # WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 # See the License for the specific language governing permissions and
 # limitations under the License.
+
 import rclpy
 from rclpy.executors import ExternalShutdownException
 from rclpy.node import Node
@@ -19,24 +20,29 @@ import rosbag2_py
 from std_msgs.msg import String
 
 
-class SimpleBagRecorder(Node):
+class CompressedBagWriter(Node):
 
     def __init__(self):
-        super().__init__('simple_bag_recorder')
-        self.writer = rosbag2_py.SequentialWriter()
+        super().__init__('compressed_bag_recorder')
+
+        compression_options = rosbag2_py.CompressionOptions(
+            compression_format='zstd',
+            compression_mode=rosbag2_py.CompressionMode.MESSAGE)
 
         storage_options = rosbag2_py.StorageOptions(
             uri='my_bag',
             storage_id='sqlite3')
         converter_options = rosbag2_py.ConverterOptions('', '')
-        self.writer.open(storage_options, converter_options)
+
+        self.compressed_writer = rosbag2_py.SequentialCompressionWriter(compression_options)
+        self.compressed_writer.open(storage_options, converter_options)
 
         topic_info = rosbag2_py.TopicMetadata(
             id=0,
             name='chatter',
             type='std_msgs/msg/String',
             serialization_format='cdr')
-        self.writer.create_topic(topic_info)
+        self.compressed_writer.create_topic(topic_info)
 
         self.subscription = self.create_subscription(
             String,
@@ -46,7 +52,7 @@ class SimpleBagRecorder(Node):
         self.subscription
 
     def topic_callback(self, msg):
-        self.writer.write(
+        self.compressed_writer.write(
             'chatter',
             serialize_message(msg),
             self.get_clock().now().nanoseconds)
@@ -55,8 +61,8 @@ class SimpleBagRecorder(Node):
 def main(args=None):
     try:
         with rclpy.init(args=args):
-            sbr = SimpleBagRecorder()
-            rclpy.spin(sbr)
+            cbr = CompressedBagWriter()
+            rclpy.spin(cbr)
     except (KeyboardInterrupt, ExternalShutdownException):
         pass
