+++ rosbag2_examples/rosbag2_examples_py/setup.py
@@ -25,11 +25,11 @@ setup(
     entry_points={
         'console_scripts': [
             'rosbag2csv = rosbag2_examples_py.rosbag2csv:main',
-            'simple_bag_recorder = rosbag2_examples_py.simple_bag_recorder:main',
+            'simple_bag_writer = rosbag2_examples_py.simple_bag_writer:main',
             'simple_bag_reader = rosbag2_examples_py.simple_bag_reader:main',
             'data_generator_node = rosbag2_examples_py.data_generator_node:main',
             'data_generator_executable = rosbag2_examples_py.data_generator_executable:main',
-            'compressed_bag_recorder = rosbag2_examples_py.compressed_bag_recorder:main',
+            'compressed_bag_writer = rosbag2_examples_py.compressed_bag_writer:main',
         ],
     },
 )
